+++ CHANGELOG.md
@@ -79,6 +79,8 @@ Main (unreleased)
   now export receivers immediately, removing the need for dependant components
   to be evaluated twice at process startup. (@rfratto)
 
+- Add missing setting to configure instance key for Eventhandler integration. (@marctc)
+
 ### Bugfixes
 
 - Remove empty port from the `apache_http` integration's instance label. (@katepangLiu)